fix: resolve sonarlint duplication error
